+++ polaris-core/src/main/java/org/apache/polaris/core/config/FeatureConfiguration.java
@@ -355,7 +355,8 @@ public class FeatureConfiguration<T> extends PolarisConfiguration<T> {
           .key("REMOTE_SIGNING_ENABLED")
           .catalogConfig("polaris.config.remote-signing.enabled")
           .description(
-              "If true, the remote signing endpoints are enabled either globally, or for a specific catalog.")
+              "If true, the remote signing endpoints are enabled either globally, or for a specific catalog. "
+                  + "This feature is currently experimental and may change in future releases.")
           .defaultValue(false)
           .buildFeatureConfiguration();