+++ polaris-core/src/main/java/org/apache/polaris/core/entity/CatalogEntity.java
@@ -26,7 +26,6 @@ import java.util.Collection;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Map;
-import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
 import org.apache.iceberg.exceptions.BadRequestException;
@@ -207,7 +206,7 @@ public class CatalogEntity extends PolarisEntity implements LocationBasedEntity 
   }
 
   public boolean isExternal() {
-    return Objects.equals(getCatalogType(), Catalog.TypeEnum.EXTERNAL);
+    return getCatalogType() == Catalog.TypeEnum.EXTERNAL;
   }
 
   public boolean isPassthroughFacade() {