+++ runtime/service/src/main/java/org/apache/polaris/service/catalog/iceberg/IcebergCatalogAdapter.java
@@ -803,7 +803,7 @@ public class IcebergCatalogAdapter
                         .addAll(PolarisEndpoints.getSupportedPolicyEndpoints(realmConfig))
                         .addAll(
                             PolarisEndpoints.getSupportedRemoteSigningEndpoints(
-                                callContext, catalogEntity))
+                                callContext.getRealmConfig(), catalogEntity))
                         .build())
                 .build())
         .build();