+++ runtime/service/src/main/java/org/apache/polaris/service/storage/aws/signer/S3RemoteSigningCatalogHandler.java
@@ -70,6 +70,8 @@ public class S3RemoteSigningCatalogHandler extends CatalogHandler implements Aut
 
     LOGGER.debug("Requesting s3 signing for {}: {}", tableIdentifier, s3SignRequest);
 
+    throwIfRemoteSigningNotEnabled(callContext.getRealmConfig(), catalogEntity);
+
     // TODO authorize based on the request's method?
 
     try {
@@ -82,8 +84,6 @@ public class S3RemoteSigningCatalogHandler extends CatalogHandler implements Aut
       authorizeCreateTableLikeUnderNamespaceOperationOrThrow(
           PolarisAuthorizableOperation.SIGN_S3_REQUEST, tableIdentifier);
     }
-
-    throwIfRemoteSigningNotEnabled(callContext.getRealmConfig(), catalogEntity);
 
     PolarisS3SignResponse s3SignResponse = s3RequestSigner.signRequest(s3SignRequest);
     LOGGER.debug("S3 signing response: {}", s3SignResponse);